refactoring, adding gmap link
