--- client/src/persons/Persons.css
@@ -1,2 +1,0 @@
-.Phones {
-}
